pyproject.toml
# ###############
# PROJECT
# ###############
[project]
name = "interloper-workspace"
version = "0.0.0"
description = "The Interloper Project"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.119.0",
    "interloper",
    "interloper-argo",
    "interloper-assets",
    "interloper-dbos",
    "interloper-docker",
    "interloper-k8s",
]

[dependency-groups]
dev = [
    "mkdocs-material>=9.6.7",
    "pre-commit>=4.2.0",
    "pyarrow>=10.0.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "python-dotenv>=1.1.1",
    "python-semantic-release>=10.0.0",
    "ruff>=0.9.4",
    "pyright>=1.1.400",
]

# ###############
# UV
# ###############
[tool.uv.sources]
interloper = { workspace = true }
interloper-argo = { workspace = true }
interloper-assets = { workspace = true }
interloper-dbos = { workspace = true }
interloper-docker = { workspace = true }
interloper-k8s = { workspace = true }

[tool.uv.workspace]
members = ["packages/*"]

[[tool.uv.index]]
name = "dc-int-connectors-prd"
url = "https://europe-python.pkg.dev/dc-int-connectors-prd/python-public/simple"
publish-url = "https://europe-python.pkg.dev/dc-int-connectors-prd/python-public"
explicit = true

# ###############
# PYRIGHT
# ###############
[tool.pyright]
include = [
    "packages/interloper/src",
    "packages/interloper-argo/src",
    "packages/interloper-assets/src",
    "packages/interloper-dbos/src",
    "packages/interloper-docker/src",
    "packages/interloper-k8s/src",
]
typeCheckingMode = "basic"
reportMissingParameterType = true

# ###############
# RUFF
# ###############
[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
extend-select = ["E", "I", "UP", "ANN001", "ANN201", "ANN202"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]
"tests/**" = ["ANN", "D102", "F811"]

# ###############
# PYTEST
# ###############
[tool.pytest.ini_options]
markers = ["integration", "functional"]
addopts = ["--import-mode=importlib", "-m not functional"]

# ###############
# COVERAGE
# ###############
[tool.coverage.report]
include = ["packages/**/src/**"]
exclude_also = [
    # Don't complain about missing debug-only code:
    # "def __repr__",
    "if self\\.debug",

    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",

    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",

    # Don't complain about abstract methods, they aren't run:
    "@(abc\\.)?abstractmethod",
]

# ################
# SEMANTIC RELEASE
# ################
[tool.semantic_release]
allow_zero_version = true
build_command = "uv build --all-packages"
version_toml = [
    "pyproject.toml:project.version",
    "packages/interloper/pyproject.toml:project.version",
]

[tool.semantic_release.branches.main]
match = "main"
prerelease = false

[tool.semantic_release.branches.rc]
match = "rc"
prerelease = true
prerelease_token = "rc"

[tool.semantic_release.branches.alpha]
match = "alpha"
prerelease = true
prerelease_token = "alpha"

[tool.semantic_release.branches.feat]
match = "feat/.*$"
prerelease = true
prerelease_token = "dev"
